Refactored database provider, moved all database initialization to databaseManager, used provider only for lifecycle management
