--- lib/widgets/auth/profile_view.dart
@@ -1,20 +1,0 @@
-import 'package:auth0_flutter/auth0_flutter.dart';
-import 'package:flutter/material.dart';
-
-class ProfileView extends StatelessWidget {
-  const ProfileView({super.key, required this.user});
-
-  final UserProfile user;
-
-  @override
-  Widget build(BuildContext context) {
-    print(user);
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: [
-        if (user.name != null) Text(user.name!),
-        if (user.email != null) Text(user.email!)
-      ],
-    );
-  }
-}
